Optional artifact logging for MLFlow (#2255)
This PR adds an option to skip artifact logging for the MlflowCallback. This is useful in cases where we want to use the only MlflowCallback for tracking metrics, and not for tracking experiment artifacts.

Additionally, this PR adds a new test to make sure that the callback works as expected when the flag is set.